% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realm_data.R
\name{read_realm}
\alias{read_realm}
\alias{read_realm}
\alias{read_realm.tbl_df}
\title{Read structure files}
\usage{
read_realm(x, ...)

\method{read_realm}{tbl_df}(x, ...)
}
\arguments{
\item{x}{structure file set}

\item{...}{unused}
}
\value{
realm
}
\description{
Read structure files
}
\examples{
library(dplyr)
files <- realm_test_files() \%>\% dplyr::filter(grepl("part\\\\.11\\\\.", fullname))
}
